Combine two subpackages and consolidate doc.go into filestystem.go for simplicity.
