+++ handler_test.go
@@ -31,7 +31,7 @@ import (
 	"github.com/pilosa/pilosa"
 	"github.com/pilosa/pilosa/internal"
 	"github.com/pilosa/pilosa/pql"
-	statik "github.com/pilosa/pilosa/statik/filesystem"
+	"github.com/pilosa/pilosa/statik"
 	"github.com/pilosa/pilosa/test"
 )
