--- statik/doc.go
@@ -1,4 +1,0 @@
-// Package statik contains static assets for the Web UI. `go generate` will
-// produce statik.go, which is ignored by git.
-//go:generate statik -src=../webui -dest=..
-package statik
